NCI-Agency/anet#3800: Add saved searches under My Work menu
