+++ client/src/pages/Home.js
@@ -7,54 +7,26 @@ import {
 } from "actions"
 import API from "api"
 import AppContext from "components/AppContext"
-import ConfirmDestructive from "components/ConfirmDestructive"
 import Fieldset from "components/Fieldset"
 import GuidedTour from "components/GuidedTour"
 import Messages from "components/Messages"
 import {
-  jumpToTop,
   mapPageDispatchersToProps,
   PageDispatchersPropType,
   useBoilerplate
 } from "components/Page"
-import SavedSearchTable from "components/SavedSearchTable"
 import { deserializeQueryParams } from "components/SearchFilters"
 import { LAST_WEEK } from "dateUtils"
-import _isEmpty from "lodash/isEmpty"
 import { Person, Report } from "models"
 import { superUserTour, userTour } from "pages/HopscotchTour"
+import MySavedSearches from "pages/searches/MySavedSearches"
 import PropTypes from "prop-types"
-import React, { useContext, useState } from "react"
-import {
-  Button,
-  ControlLabel,
-  FormControl,
-  FormGroup,
-  Grid,
-  Row
-} from "react-bootstrap"
+import React, { useContext } from "react"
+import { Button, Grid, Row } from "react-bootstrap"
 import { connect } from "react-redux"
 import { useHistory, useLocation } from "react-router-dom"
 import { RECURSE_STRATEGY } from "searchUtils"
 import Settings from "settings"
-import utils from "utils"
-
-const GQL_GET_SAVED_SEARCHES = gql`
-  query {
-    savedSearches: mySearches {
-      uuid
-      name
-      objectType
-      query
-    }
-  }
-`
-
-const GQL_DELETE_SAVED_SEARCH = gql`
-  mutation($uuid: String!) {
-    deleteSavedSearch(uuid: $uuid)
-  }
-`
 
 const GQL_GET_REPORT_COUNT = gql`
   query($reportQuery: ReportSearchQueryInput) {
@@ -296,112 +268,6 @@ HomeTiles.propTypes = {
   currentUser: PropTypes.instanceOf(Person)
 }
 
-const SavedSearches = ({ setSearchQuery, pageDispatchers }) => {
-  const history = useHistory()
-  const [stateError, setStateError] = useState(null)
-  const [selectedSearch, setSelectedSearch] = useState(null)
-  const { loading, error, data, refetch } = API.useApiQuery(
-    GQL_GET_SAVED_SEARCHES
-  )
-  const { done, result } = useBoilerplate({
-    loading,
-    error,
-    pageDispatchers
-  })
-  if (done) {
-    return result
-  }
-
-  let savedSearches = []
-  if (data) {
-    savedSearches = data.savedSearches
-    if (_isEmpty(savedSearches)) {
-      if (selectedSearch) {
-        // Clear selection
-        setSelectedSearch(null)
-      }
-    } else if (!savedSearches.includes(selectedSearch)) {
-      // Select first one
-      setSelectedSearch(savedSearches[0])
-    }
-  }
-
-  return (
-    <>
-      <Messages error={stateError} />
-      <FormGroup controlId="savedSearchSelect">
-        <ControlLabel>Select a saved search</ControlLabel>
-        <FormControl componentClass="select" onChange={onSaveSearchSelect}>
-          {savedSearches &&
-            savedSearches.map(savedSearch => (
-              <option value={savedSearch.uuid} key={savedSearch.uuid}>
-                {savedSearch.name}
-              </option>
-            ))}
-        </FormControl>
-      </FormGroup>
-
-      {selectedSearch && (
-        <div>
-          <div className="pull-right">
-            <Button style={{ marginRight: 12 }} onClick={showSearch}>
-              Show Search
-            </Button>
-            <ConfirmDestructive
-              onConfirm={onConfirmDelete}
-              objectType="search"
-              objectDisplay={selectedSearch.name}
-              bsStyle="danger"
-              buttonLabel="Delete Search"
-            />
-          </div>
-          <SavedSearchTable search={selectedSearch} />
-        </div>
-      )}
-    </>
-  )
-
-  function onSaveSearchSelect(event) {
-    const uuid = event && event.target ? event.target.value : event
-    const search = savedSearches.find(el => el.uuid === uuid)
-    setSelectedSearch(search)
-  }
-
-  function showSearch() {
-    if (selectedSearch) {
-      const objType = SEARCH_OBJECT_TYPES[selectedSearch.objectType]
-      const queryParams = utils.parseJsonSafe(selectedSearch.query)
-      deserializeQueryParams(objType, queryParams, deserializeCallback)
-    }
-  }
-
-  function deserializeCallback(objectType, filters, text) {
-    // We update the Redux state
-    setSearchQuery({
-      objectType: objectType,
-      filters: filters,
-      text: text
-    })
-    history.push("/search")
-  }
-
-  function onConfirmDelete() {
-    return API.mutation(GQL_DELETE_SAVED_SEARCH, { uuid: selectedSearch.uuid })
-      .then(data => {
-        refetch()
-      })
-      .catch(error => {
-        setStateError(error)
-        jumpToTop()
-      })
-  }
-}
-
-SavedSearches.propTypes = {
-  setSearchQuery: PropTypes.func.isRequired,
-  pageDispatchers: PageDispatchersPropType
-}
-
 const Home = ({ setSearchQuery, pageDispatchers }) => {
   const { currentUser } = useContext(AppContext)
   const routerLocation = useLocation()
@@ -464,12 +330,10 @@ const Home = ({ setSearchQuery, pageDispatchers }) => {
         />
       </Fieldset>
 
-      <Fieldset title="Saved searches">
-        <SavedSearches
-          setSearchQuery={setSearchQuery}
-          pageDispatchers={pageDispatchers}
-        />
-      </Fieldset>
+      <MySavedSearches
+        setSearchQuery={setSearchQuery}
+        pageDispatchers={pageDispatchers}
+      />
     </div>
   )
 }
